[refactor]packaging controller codes into core
