+++ tests/test_integration_bw_controller_and_shell.py
@@ -1,0 +1,7 @@
+from unittest.mock import mock_open
+from pytest_mock import mocker, MockerFixture
+from shell.shell import Shell
+import pytest
+
+def test_read():
+    ...
